Some file restructuring.
